Using __op_add for addition instead of BinaryOperator.
This is so that we can add numbers of different types. An end-to-end
test was also added to do addition with implicit casting:
add_with_implicit_casting.alusus

+++ Sources/Scg/BuiltInFunctions/AddFloats.cpp
@@ -23,8 +23,8 @@ namespace Scg
 {
 AddFloats::AddFloats()
 {
-  this->argTypeSpecs.push_back(new ValueTypeSpecByName("int"));
-  this->argTypeSpecs.push_back(new ValueTypeSpecByName("int"));
+  this->argTypeSpecs.push_back(new ValueTypeSpecByName("float"));
+  this->argTypeSpecs.push_back(new ValueTypeSpecByName("float"));
 }
 
 AddFloats::~AddFloats()
